% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_apply_model.R
\name{apply_model}
\alias{apply_model}
\title{Apply Model}
\usage{
apply_model(
  obj,
  models = c("SL.glm", "SL.svm"),
  test_partition_prop = 0.2,
  perf_measure = "auc"
)
}
\arguments{
\item{obj}{object returned from 02_define_depvars_outcome}

\item{models}{models to be analyzed}

\item{test_partition_prop}{test proportion}

\item{perf_measure}{performance measure}
}
\value{
list
}
\description{
Apply model and create column with fit
}
\examples{
apply_model(obj2)
}
