% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bananaquality}
\alias{bananaquality}
\title{Banana Quality}
\format{
An object of class \code{data.frame} with 8000 rows and 8 columns.
}
\usage{
bananaquality
}
\description{
Banana quality dataset
}
\keyword{datasets}
