% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bananaquality_sample}
\alias{bananaquality_sample}
\title{Banana Quality Subset}
\format{
An object of class \code{data.frame} with 50 rows and 8 columns.
}
\usage{
bananaquality_sample
}
\description{
Banana quality dataset subset
}
\keyword{datasets}
