% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_bootstrap_data.R
\name{bootstrap_data}
\alias{bootstrap_data}
\title{Bootstrap data}
\usage{
bootstrap_data(
  data = bananaquality[sample(1:nrow(bananaquality), replace = FALSE, size = 200), ],
  outcome = "Quality",
  indep_vars = setdiff(names(bananaquality), outcome),
  n_samples = 50,
  n_maximum_dim = 5
)
}
\arguments{
\item{data}{a dataframe to be analysed}

\item{outcome}{a string representing the outcome variable}

\item{indep_vars}{a vector of strings to be considered}

\item{n_samples}{number of bootstrap samples}

\item{n_maximum_dim}{maximum number of variables to be considered}
}
\value{
list
}
\description{
Create a tibble with a number of bootstrap samples
}
\examples{
bootstrap_data()
}
