% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obj2}
\alias{obj2}
\title{Example Object returned from \code{define_indepvars_outcome}}
\format{
An object of class \code{list} of length 7.
}
\usage{
obj2
}
\description{
obj2
}
\keyword{datasets}
