% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_curve.R
\name{plot_curve}
\alias{plot_curve}
\title{Plot curve}
\usage{
plot_curve(obj)
}
\arguments{
\item{obj}{object returned by \code{summary_statistics}}
}
\value{
list with graphical features
}
\description{
Return plot features.
}
\examples{
plot_curve(obj4)
}
