% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_summary_statistics.R
\name{summary_stats}
\alias{summary_stats}
\title{Summary Stats}
\usage{
summary_stats(obj)
}
\arguments{
\item{obj}{object returned from  \code{apply_model}}
}
\value{
list with summary statistics and bootstrap confidence intervals
}
\description{
Return summary statistics
}
\examples{
summary_stats(obj3)
}
