% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aucdim.R
\name{wrapper_aucdim}
\alias{wrapper_aucdim}
\title{Wrap all pipeline}
\usage{
wrapper_aucdim(
  data,
  outcome,
  indep_vars,
  n_samples = 100,
  n_maximum_dim = 5,
  p_in = 0.5,
  p_out = 0.6,
  models = c("SL.glm"),
  test_partition_prop = 0.2,
  perf_measure = "auc",
  x_label_offset = 1,
  y_label_offset = 10
)
}
\arguments{
\item{data}{a dataframe to be analyzed}

\item{outcome}{a string representing the outcome variable}

\item{indep_vars}{a vector of strings to be considered}

\item{n_samples}{number of bootstrap samples}

\item{n_maximum_dim}{maximum number of variables}

\item{p_in}{entry p-value for choosing variable order}

\item{p_out}{exclusion p-value for choosing variable order}

\item{models}{a string representing the models to compare}

\item{test_partition_prop}{test partition proportion}

\item{perf_measure}{performance measure to be considered}

\item{x_label_offset}{x coordinate for plotting}

\item{y_label_offset}{y coordinate for plotting}
}
\value{
a list with the final object
}
\description{
Wrap all pipeline
}
