% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_bootstrap_data.R
\name{bootstrap_data}
\alias{bootstrap_data}
\title{Bootstrap data}
\usage{
bootstrap_data(
  data,
  outcome = "Quality",
  indep_vars = c("Size", "Weight", "Sweetness", "Softness", "HarvestTime", "Ripeness",
    "Acidity"),
  n_samples = 50,
  n_maximum_dim = 5
)
}
\arguments{
\item{data}{a dataframe to be analyzed}

\item{outcome}{a string representing the outcome variable}

\item{indep_vars}{a vector of strings to be considered}

\item{n_samples}{number of bootstrap samples}

\item{n_maximum_dim}{maximum number of variables to be considered}
}
\value{
list
}
\description{
Create a list with bootstrap samples
}
\examples{
bootstrap_data(bananaquality_sample)
}
