% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_compare_test.R
\name{compare_test}
\alias{compare_test}
\title{Comapare test}
\usage{
compare_test(obj)
}
\arguments{
\item{obj}{object returned by plot_curve}
}
\value{
list
}
\description{
Comapare test
}
\examples{
compare_test(obj)
}
