% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_define_depvars_outcome.R
\name{define_indepvars}
\alias{define_indepvars}
\title{Define dep vars}
\usage{
define_indepvars(obj, p_in = 0.5, p_out = 0.6)
}
\arguments{
\item{obj}{object returned by bootstrap_data}

\item{p_in}{entry p-value used to determine variable order}

\item{p_out}{removal p-value used to determine variable order}
}
\value{
list
}
\description{
Create a column with dependent variables
}
\examples{
define_indepvars(obj)
}
