% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_summary_statistics.R
\name{summary_stats}
\alias{summary_stats}
\title{Summary Stats}
\usage{
summary_stats(obj)
}
\arguments{
\item{obj}{object returned by apply_model}
}
\value{
list
}
\description{
Summary Stats
}
\examples{
summary_stats(obj3)
}
