% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aucdim.R
\name{wrapper_aucdim}
\alias{wrapper_aucdim}
\title{Wrap all pipeline}
\usage{
wrapper_aucdim(data, outcome, indep_vars, ...)
}
\arguments{
\item{data}{a dataframe to be analysed}

\item{outcome}{a string representing the outcome variable}

\item{indep_vars}{a vector of strings to be considered}

\item{...}{arguments used across pipeline}
}
\value{
tibble
}
\description{
Wrap all pipeline
}
