% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aucdim.R
\name{wrapper_aucdim}
\alias{wrapper_aucdim}
\title{Wrap all pipeline}
\usage{
wrapper_aucdim(
  data,
  outcome,
  indep_vars,
  n_samples = 100,
  n_maximum_dim = 5,
  p_in = 0.5,
  p_out = 0.6,
  models = c("SL.glm"),
  test_partition_prop = 0.2,
  perf_measure = "auc",
  x_label_offset = 1,
  y_label_offset = 10
)
}
\arguments{
\item{data}{a dataframe to be analyzed}

\item{outcome}{a string representing the outcome variable}

\item{indep_vars}{a vector of strings to be considered}

\item{...}{arguments used across pipeline}
}
\value{
a list with the final object
}
\description{
Wrap all pipeline
}
